+++ state/src/lib.rs
@@ -1,8 +1,8 @@
 mod cursor;
 mod custom_attributes;
 mod font_style;
-mod references;
 mod layout;
+mod references;
 mod style;
 mod transform;
 mod values;
@@ -10,8 +10,8 @@ mod values;
 pub use cursor::*;
 pub use custom_attributes::*;
 pub use font_style::*;
-pub use references::*;
 pub use layout::*;
+pub use references::*;
 pub use style::*;
 pub use transform::*;
 pub use values::*;
