+++ state/src/values/decoration.rs
@@ -1,5 +1,5 @@
-use skia_safe::textlayout::{TextDecoration, TextDecorationStyle};
 use crate::Parse;
+use skia_safe::textlayout::{TextDecoration, TextDecorationStyle};
 
 #[derive(Debug, PartialEq, Eq)]
 pub struct ParseTextDecorationError;
@@ -8,22 +8,22 @@ impl Parse for TextDecoration {
     type Err = ParseTextDecorationError;
 
     fn parse(value: &str) -> Result<Self, Self::Err> {
-		let mut decoration = TextDecoration::default();
-		let values = value.split_ascii_whitespace();
-	
-		for val in values {
-			decoration.set(
-				match val {
-					"underline" => TextDecoration::UNDERLINE,
-					"overline" => TextDecoration::OVERLINE,
-					"line-through" => TextDecoration::LINE_THROUGH,
-					_ => TextDecoration::NO_DECORATION,
-				},
-				true,
-			);
-		}
+        let mut decoration = TextDecoration::default();
+        let values = value.split_ascii_whitespace();
 
-		Ok(decoration)
+        for val in values {
+            decoration.set(
+                match val {
+                    "underline" => TextDecoration::UNDERLINE,
+                    "overline" => TextDecoration::OVERLINE,
+                    "line-through" => TextDecoration::LINE_THROUGH,
+                    _ => TextDecoration::NO_DECORATION,
+                },
+                true,
+            );
+        }
+
+        Ok(decoration)
     }
 }
 
@@ -34,13 +34,13 @@ impl Parse for TextDecorationStyle {
     type Err = ParseTextDecorationStyleError;
 
     fn parse(value: &str) -> Result<Self, Self::Err> {
-		Ok(match value {
-			"solid" => TextDecorationStyle::Solid,
-			"double" => TextDecorationStyle::Double,
-			"dotted" => TextDecorationStyle::Dotted,
-			"dashed" => TextDecorationStyle::Dashed,
-			"wavy" => TextDecorationStyle::Wavy,
-			_ => TextDecorationStyle::Solid,
-		})
+        Ok(match value {
+            "solid" => TextDecorationStyle::Solid,
+            "double" => TextDecorationStyle::Double,
+            "dotted" => TextDecorationStyle::Dotted,
+            "dashed" => TextDecorationStyle::Dashed,
+            "wavy" => TextDecorationStyle::Wavy,
+            _ => TextDecorationStyle::Solid,
+        })
     }
 }
