+++ state/src/values/display.rs
@@ -1,5 +1,5 @@
-use torin::display::DisplayMode;
 use crate::Parse;
+use torin::display::DisplayMode;
 
 #[derive(Debug, PartialEq, Eq)]
 pub struct ParseDisplayModeError;
@@ -8,9 +8,9 @@ impl Parse for DisplayMode {
     type Err = ParseDisplayModeError;
 
     fn parse(value: &str) -> Result<Self, Self::Err> {
-		Ok(match value {
-			"center" => DisplayMode::Center,
-			_ => DisplayMode::Normal,
-		})
+        Ok(match value {
+            "center" => DisplayMode::Center,
+            _ => DisplayMode::Normal,
+        })
     }
 }
