+++ state/tests/parse_shadows.rs
@@ -1,4 +1,4 @@
-use freya_node_state::{Parse, Shadow};
+use freya_node_state::{Parse, Shadow, ShadowPosition};
 use skia_safe::Color;
 
 #[test]
@@ -12,7 +12,7 @@ fn parse_big_shadow() {
             blur: 50.0,
             spread: 25.0,
             color: Color::RED,
-            inset: false
+            position: ShadowPosition::Inset
         })
     );
 }
@@ -28,7 +28,7 @@ fn parse_inset_shadow() {
             blur: 50.0,
             spread: 25.0,
             color: Color::RED,
-            inset: true
+            position: ShadowPosition::Inset
         })
     );
 }
@@ -44,7 +44,7 @@ fn parse_shadow_with_assumed_spread() {
             blur: 50.0,
             spread: 0.0,
             color: Color::RED,
-            inset: true
+            position: ShadowPosition::Inset
         })
     );
 }
