+++ torin/src/lib.rs
@@ -1,24 +1,19 @@
 pub mod custom_measurer;
-pub mod direction;
-pub mod display;
 pub mod dom_adapter;
 pub mod geometry;
 pub mod node;
-pub mod padding;
-pub mod radius;
-pub mod size;
 pub mod scaled;
 pub mod torin;
+pub mod values;
+
+pub use values::*;
 
 pub mod prelude {
     pub use crate::custom_measurer::*;
-    pub use crate::direction::*;
-    pub use crate::display::*;
     pub use crate::dom_adapter::*;
     pub use crate::geometry::*;
     pub use crate::node::*;
-    pub use crate::padding::*;
-    pub use crate::size::*;
-    pub use crate::torin::*;
     pub use crate::scaled::*;
+    pub use crate::torin::*;
+    pub use crate::values::prelude::*;
 }
